requirements.txt
asgiref==3.9.1
django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
pyjwt==2.10.1
sqlparse==0.5.3
